Many changes were made read description
Files can now be uploaded to the server using npm package for middleware ('express-fileupload').  For now, just the name of the file received by server is being printed since I'm not familiar with how to send a file to an ip address. Also the devices endpoint is now compatible with the object being received.