Resolved audio feedback issue
Previously the user's own audio would output on their machine as well causing a feedback loop. Using the mute button would also mute the audio for other users on the call. Solution was to mute audio output only on the video element on the user's room but not source object of the video stream.